requirements.txt
fastapi
uvicorn
pyyaml
google-generativeai
anthropic
